requirements.txt
Faker==21.0.0
pika==1.3.2
python-dotenv 1.0.0
python-dateutil
six
